Consolidar 4 tests DT/DTM en archivos por modulo (-278 lineas netas)
- test_dt6_context_truncation.py + test_dt9_extract_score_fallback.py
  → absorbidos en test_metrics_reference_based.py (pass-through + 21 casos regex)
- test_dt8_09_10_11_reranker_sort.py → TestRerankSortEdges en test_reranker.py
  (identical scores, missing relevance_score)
- test_dtm4_loader_populate.py → test_loader.py
  (10 casos edge de _populate_from_dataframes)

Bloque E18 parcialmente cerrado. TESTS.md actualizado.
